Divide project e2e tests in three files
